+++ Task026.cs
@@ -6,7 +6,7 @@ namespace euler_from26
 {
     public static class Task26
     {
-        public static int cycle_len(int p)
+        public static long cycle_len(long p)
         {
             for (int candidate = 2; candidate <= p; candidate++)
                 if (BigInteger.ModPow(10, candidate, p) == 1) 
@@ -16,11 +16,11 @@ namespace euler_from26
 
         public static void main()
         {
-            int max = 0;
-            int max_d = 0;
+            long max = 0;
+            long max_d = 0;
             foreach(var d in Prime.Prime_to(1000))
             {
-                int c = cycle_len(d);
+                long c = cycle_len(d);
                 if (c > max)
                 {
                     max = c;
