+++ Task033.cs
@@ -7,8 +7,8 @@ namespace euler_from26
     {
         public static void main()
         {
-            for(int numerator = 10; numerator < 99; numerator++)
-                for (int denominator = numerator + 1; denominator < 100; denominator++)
+            for(long numerator = 10; numerator < 99; numerator++)
+                for (long denominator = numerator + 1; denominator < 100; denominator++)
                 {
                     var n_d = Digits.digits_list(numerator);
                     var d_d = Digits.digits_list(denominator);
