+++ Task035.cs
@@ -6,8 +6,8 @@ namespace euler_from26
     public static class Task35
     {
 
-        public static int[] primes;
-        public static bool isCircular(List<int> digits)
+        public static long[] primes;
+        public static bool isCircular(List<long> digits)
         {
             for (int i = 0; i < digits.Count - 1; i++)
             {
